+++ CleanApp/DataTests/Helpers/AccountModelFactory.swift
@@ -1,0 +1,18 @@
+//
+//  AccountModelFactory.swift
+//  DataTests
+//
+//  Created by Alexandre Cardoso on 20/04/22.
+//
+
+import Foundation
+import Domain
+
+func makeAccountModel() -> AccountModel {
+    return AccountModel(
+        id: "any_id",
+        name: "any_name",
+        email: "any_email@mail.com",
+        password: "any_password"
+    )
+}
